DBZ-252 - extract creation of DdlParser interface
