+++ debezium-core/src/test/java/io/debezium/relational/ddl/DdlParserSql2003Test.java
@@ -19,7 +19,7 @@ import io.debezium.relational.Tables;
 
 public class DdlParserSql2003Test {
 
-    private DdlParser parser;
+    private LegacyDdlParser parser;
     private Tables tables;
 
     @Before
